% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KDB.R
\name{cdf.KDB}
\alias{cdf.KDB}
\title{Evaluate the cumulative distribution function of a KDB distribution}
\usage{
\method{cdf}{KDB}(d, x, ...)
}
\arguments{
\item{d}{A `KDB` object created by a call to [KDB()].}

\item{x}{A vector of elements whose cumulative probabilities you would
like to determine given the distribution `d`.}

\item{...}{Unused. Unevaluated arguments will generate a warning to
catch mispellings or other possible errors.}
}
\value{
A vector of probabilities, one for each element of `x`.
}
\description{
Evaluate the cumulative distribution function of a KDB distribution
}
\examples{

set.seed(27)

X <- KDB(alpha=2,beta=5)
X

random(X, 10)

pdf(X, 0.7)
log_pdf(X, 0.7)

cdf(X, 0.7)
quantile(X, 0.7)

cdf(X, quantile(X, 0.7))
quantile(X, cdf(X, 0.7))
}
