% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Triangular.R
\name{cdf.Triangular}
\alias{cdf.Triangular}
\title{Evaluate the cumulative distribution function of a Triangular distribution}
\usage{
\method{cdf}{Triangular}(d, x, ...)
}
\arguments{
\item{d}{A `Triangular` object created by a call to [Triangular()].}

\item{x}{A vector of elements whose cumulative probabilities you would
like to determine given the distribution `d`.}

\item{...}{Unused. Unevaluated arguments will generate a warning to
catch mispellings or other possible errors.}
}
\value{
A vector of probabilities, one for each element of `x`.
}
\description{
Evaluate the cumulative distribution function of a Triangular distribution
}
\examples{

set.seed(27)

X <- Triangular(a=0,b=4,c=1)
X

random(X, 10)

pdf(X, 0.7)
log_pdf(X, 0.7)

cdf(X, 0.7)
quantile(X, 0.7)

cdf(X, quantile(X, 0.7))
quantile(X, cdf(X, 0.7))
}
