% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpqrMethods.R
\name{cdf_trim}
\alias{cdf_trim}
\title{Cdf of a trimmed distribution}
\usage{
cdf_trim(d, x, trim = c(-Inf, Inf), doTrunc = FALSE)
}
\arguments{
\item{d}{distributions3 object}

\item{x}{Numeric vector, values at which the cdf is evaluated}

\item{trim}{Numeric vector of size 2, trimming bounds.}

\item{doTrunc}{Logical, do truncation? default FALSE, i.e. rectification is used.}
}
\value{
a numeric vector containing the cdf values.
}
\description{
Evaluate the cdf of a trimmed distribution.
}
\examples{
# Define Normal distribution
norm <- Normal(mu=0.75,sigma=0.5)
# Define a grid of values x
x=seq(-1,2,,100)
# Compare cdf of basic / rectified / truncated Gaussian distributions
y0=cdf_trim(norm,x)
yr=cdf_trim(norm,x,trim=c(0,1))
yt=cdf_trim(norm,x,trim=c(0,1),doTrunc=TRUE)
plot(x,y0,type='l',ylim=c(0,1))
points(x,yr,col='red')
points(x,yt,col='blue',pch=3)
}
