% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions3_utils.R
\name{copyDist}
\alias{copyDist}
\title{copyDist}
\usage{
copyDist(d, param = getParVector(d))
}
\arguments{
\item{d}{distributions3 object to be copied (provides the distribution family)}

\item{param}{numeric vector of size npar(d), new parameter vector}
}
\value{
a distributions3 object.
}
\description{
Make a copy of a distributions3 object and reset its parameter vector.
}
\examples{
\dontrun{
d1 <- Normal()
d2 <- copyDist(d,param=c(10,5))
}
}
\keyword{internal}
