% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CowhouseCreek}
\alias{CowhouseCreek}
\title{Cowhouse Creek low flow dataset}
\format{
An object of class \code{data.frame} with 69 rows and 3 columns.
}
\source{
\url{https://waterdata.usgs.gov/nwis/inventory/?site_no=08101000&agency_cd=USGS}
}
\usage{
CowhouseCreek
}
\description{
Annual minimum streamflow (in m3/s) and annual duration below
0.01 m3/s (in year) recorded at the Cowhouse Creek station at
Pidcoke, Texas, USA, 1951-2019
}
\keyword{datasets}
