% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{feas}
\alias{feas}
\title{Check feasability of a distribution (e.g. positive scale,
lower / upper bounds in correct order, etc.)}
\usage{
feas(d)
}
\arguments{
\item{d}{distributions3 object}
}
\value{
a logical, is the distribution feasible?
}
\description{
Check feasability of a distribution (e.g. positive scale,
lower / upper bounds in correct order, etc.)
}
\examples{
feas(Triangular())
feas(Triangular(a=2,b=1))
}
