% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions3_utils.R
\name{getParName}
\alias{getParName}
\title{Get parameter names}
\usage{
getParName(d)
}
\arguments{
\item{d}{distributions3 object}
}
\value{
a character vector containing the parameter names of d.
}
\description{
Get parameter names from a distributions3 object.
}
\examples{
\dontrun{
d <- Normal()
getParName(d)
}
}
\keyword{internal}
