% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions3_utils.R
\name{getParVector}
\alias{getParVector}
\title{Get parameter vector}
\usage{
getParVector(d)
}
\arguments{
\item{d}{distributions3 object}
}
\value{
a numeric vector containing the parameters of d.
}
\description{
Get parameter vector from a distributions3 object.
Note that this function is highly dependent to the
structure of a distributions3 object. If one day the developpers decide
to add more fields to this structure, this function will have to be changed!
}
\examples{
\dontrun{
d <- Normal()
getParVector(d)
}
}
\keyword{internal}
