% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getPs}
\alias{getPs}
\title{Get left / middle / right probabilities}
\usage{
getPs(d, trim = c(-Inf, Inf))
}
\arguments{
\item{d}{distributions3 object}

\item{trim}{Numeric vector of size 2, trimming bounds.}
}
\value{
a list with values left,right,mid.
}
\description{
Get left / middle / right probabilities
}
\keyword{internal}
