% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEV.R
\name{GEV}
\alias{GEV}
\title{Create a GEV distribution (Generalized Extreme Value)}
\usage{
GEV(loc = 0, scale = 1, shape = 0)
}
\arguments{
\item{loc}{The location parameter, often noted \eqn{\mu}. Can be any real number.
Defaults to `0`.}

\item{scale}{The scale parameter, often noted \eqn{\sigma}. Can be any
positive number. Defaults to `1`.}

\item{shape}{The shape parameter, often noted \eqn{\xi}. Can be any
real number. Defaults to `0`, in which case the GEV distribution
is also known as the Gumbel distribution. Note that the 'statistical'
convention (also used in package evd, in Wikipedia and in many textbooks)
is used here: \eqn{\xi > 0} correspond to heavy-tailed, left-bounded
distributions (aka Frechet Family or Type-II distributions).
\eqn{\xi < 0} correspond to light-tailed, right-bounded
distributions (aka Weibull Family or Type-III distributions). Be aware that
the opposite convention is used in some disciplines (typically, Hydrology).}
}
\value{
A `GEV` object.
}
\description{
Create a GEV distribution (Generalized Extreme Value)
}
\examples{

set.seed(27)

X <- GEV(loc=0,scale=1,shape=0.2)
X

random(X, 10)

pdf(X, 0.7)
log_pdf(X, 0.7)

cdf(X, 0.7)
quantile(X, 0.7)

cdf(X, quantile(X, 0.7))
quantile(X, cdf(X, 0.7))
}
\seealso{
Other continuous distributions: 
\code{\link{GEVmin2}()},
\code{\link{GEVmin}()},
\code{\link{GPDmin2}()},
\code{\link{GPDmin}()},
\code{\link{GPD}()},
\code{\link{KDB4}()},
\code{\link{KDB}()},
\code{\link{Triangular}()}
}
\concept{continuous distributions}
