% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPD.R
\name{GPD}
\alias{GPD}
\title{Create a GPD distribution (Generalized Pareto Distribution)}
\usage{
GPD(loc = 0, scale = 1, shape = 0)
}
\arguments{
\item{loc}{The location parameter, often noted \eqn{\mu}, which represents
the threshold above which data are taken. Can be any real number. Defaults to `0`.}

\item{scale}{The scale parameter, often noted \eqn{\sigma}. Can be any
positive number. Defaults to `1`.}

\item{shape}{The shape parameter, often noted \eqn{\xi}. Can be any
real number. Defaults to `0`, in which case the GPD distribution
correspond to the Exponential distribution. Note that the 'statistical'
convention (also used in package evd, in Wikipedia and in many textbooks)
is used here: \eqn{\xi > 0} correspond to heavy-tailed distributions.
\eqn{\xi < 0} correspond to light-tailed, right-bounded distributions.
Be aware that the opposite convention is used in some disciplines (typically, Hydrology).}
}
\value{
A `GPD` object.
}
\description{
Create a GPD distribution (Generalized Pareto Distribution)
}
\examples{

set.seed(27)

X <- GPD(loc=0,scale=1,shape=0.2)
X

random(X, 10)

pdf(X, 0.7)
log_pdf(X, 0.7)

cdf(X, 0.7)
quantile(X, 0.7)

cdf(X, quantile(X, 0.7))
quantile(X, cdf(X, 0.7))
}
\seealso{
Other continuous distributions: 
\code{\link{GEVmin2}()},
\code{\link{GEVmin}()},
\code{\link{GEV}()},
\code{\link{GPDmin2}()},
\code{\link{GPDmin}()},
\code{\link{KDB4}()},
\code{\link{KDB}()},
\code{\link{Triangular}()}
}
\concept{continuous distributions}
