% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPDmin.R
\name{GPDmin}
\alias{GPDmin}
\title{Create a GPDmin distribution (Generalized Pareto Distribution for minima)}
\usage{
GPDmin(loc = 0, scale = 1, shape = 0)
}
\arguments{
\item{loc}{The location parameter, often noted \eqn{\mu}, which represents
the threshold below which data are taken. Can be any real number. Defaults to `0`.}

\item{scale}{The scale parameter, often noted \eqn{\sigma}. Can be any
positive number. Defaults to `1`.}

\item{shape}{The shape parameter, often noted \eqn{\xi}. Can be any
real number. Defaults to `0`. Note that the convention used here derives from
the one used for the GPD distribution:
\eqn{\xi > 0} leads to a right-bounded distribution.
\eqn{\xi < 0} leads to a left-bounded distribution.}
}
\value{
A `GPDmin` object.
}
\description{
Create a GPDmin distribution (Generalized Pareto Distribution for minima)
}
\examples{

set.seed(27)

X <- GPDmin(loc=1,scale=1,shape=-0.2)
X

random(X, 10)

pdf(X, 0.7)
log_pdf(X, 0.7)

cdf(X, 0.7)
quantile(X, 0.7)

cdf(X, quantile(X, 0.7))
quantile(X, cdf(X, 0.7))
}
\seealso{
Other continuous distributions: 
\code{\link{GEVmin2}()},
\code{\link{GEVmin}()},
\code{\link{GEV}()},
\code{\link{GPDmin2}()},
\code{\link{GPD}()},
\code{\link{KDB4}()},
\code{\link{KDB}()},
\code{\link{Triangular}()}
}
\concept{continuous distributions}
