% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPDmin2.R
\name{GPDmin2}
\alias{GPDmin2}
\title{Create a GPDmin2 distribution (Generalized Pareto Distribution for minima with alternative parameterization)}
\usage{
GPDmin2(loc = 0, scale = 1, lbound = -1)
}
\arguments{
\item{loc}{The location parameter, often noted \eqn{\mu}, which represents
the threshold below which data are taken. Can be any real number.
Defaults to `0`.}

\item{scale}{The scale parameter, often noted \eqn{\sigma}. Can be any
positive number. Defaults to `1`.}

\item{lbound}{The lower bound parameter \eqn{\alpha}. Can be any
real number. Defaults to `-1`. It should be strictly smaller than the location
parameter \eqn{\mu}. Relation with the standard GPD parameterization is
\eqn{\alpha=\mu+\sigma/\xi} and \eqn{\xi<0}.}
}
\value{
A `GPDmin2` object.
}
\description{
Create a GPDmin2 distribution (Generalized Pareto Distribution for minima with alternative parameterization)
}
\examples{

set.seed(27)

X <- GPDmin2(loc=2,scale=1,lbound=0)
X

random(X, 10)

pdf(X, 0.7)
log_pdf(X, 0.7)

cdf(X, 0.7)
quantile(X, 0.7)

cdf(X, quantile(X, 0.7))
quantile(X, cdf(X, 0.7))
}
\seealso{
Other continuous distributions: 
\code{\link{GEVmin2}()},
\code{\link{GEVmin}()},
\code{\link{GEV}()},
\code{\link{GPDmin}()},
\code{\link{GPD}()},
\code{\link{KDB4}()},
\code{\link{KDB}()},
\code{\link{Triangular}()}
}
\concept{continuous distributions}
