% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KDB.R
\name{KDB}
\alias{KDB}
\title{Create a KDB distribution (Kuramaswamy double-bounded)}
\usage{
KDB(alpha = 1, beta = 1)
}
\arguments{
\item{alpha}{The alpha parameter (first shape parameter).
`alpha` can be any value strictly greater than zero. Defaults to `1`.}

\item{beta}{The beta parameter (second shape parameter).
`beta` can be any value strictly greater than zero. Defaults to `1`.}
}
\value{
A `KDB` object.
}
\description{
Create a KDB distribution (Kuramaswamy double-bounded)
}
\examples{

set.seed(27)

X <- KDB(alpha=2,beta=5)
X

random(X, 10)

pdf(X, 0.7)
log_pdf(X, 0.7)

cdf(X, 0.7)
quantile(X, 0.7)

cdf(X, quantile(X, 0.7))
quantile(X, cdf(X, 0.7))
}
\seealso{
Other continuous distributions: 
\code{\link{GEVmin2}()},
\code{\link{GEVmin}()},
\code{\link{GEV}()},
\code{\link{GPDmin2}()},
\code{\link{GPDmin}()},
\code{\link{GPD}()},
\code{\link{KDB4}()},
\code{\link{Triangular}()}
}
\concept{continuous distributions}
