% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KDB4.R
\name{KDB4}
\alias{KDB4}
\title{Create a KDB4 distribution (4-parameter Kuramaswamy double-bounded)}
\usage{
KDB4(alpha = 1, beta = 1, a = 0, b = 1)
}
\arguments{
\item{alpha}{The alpha parameter (first shape parameter).
`alpha` can be any value strictly greater than zero. Defaults to `1`.}

\item{beta}{The beta parameter (second shape parameter).
`beta` can be any value strictly greater than zero. Defaults to `1`.}

\item{a}{The a parameter (lower bound). `a` can be any value in the set of real
numbers. Defaults to `0`.}

\item{b}{The b parameter (upper bound). `b` can be any value in the set of real
numbers. It should be strictly bigger than `a`. Defaults to `1`.}
}
\value{
A `KDB4` object.
}
\description{
Create a KDB4 distribution (4-parameter Kuramaswamy double-bounded)
}
\examples{

set.seed(27)

X <- KDB4(alpha=2,beta=5,a=0.5,b=3)
X

random(X, 10)

pdf(X, 0.7)
log_pdf(X, 0.7)

cdf(X, 0.7)
quantile(X, 0.7)

cdf(X, quantile(X, 0.7))
quantile(X, cdf(X, 0.7))
}
\seealso{
Other continuous distributions: 
\code{\link{GEVmin2}()},
\code{\link{GEVmin}()},
\code{\link{GEV}()},
\code{\link{GPDmin2}()},
\code{\link{GPDmin}()},
\code{\link{GPD}()},
\code{\link{KDB}()},
\code{\link{Triangular}()}
}
\concept{continuous distributions}
