% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions3_utils.R
\name{nPar}
\alias{nPar}
\title{Get number of parameters}
\usage{
nPar(d)
}
\arguments{
\item{d}{distributions3 object}
}
\value{
the number of parameters of d.
}
\description{
Get number of parameters
}
\examples{
\dontrun{
d <- Normal()
nPar(d)
}
}
\keyword{internal}
