% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEVmin2.R
\name{pdf.GEVmin2}
\alias{pdf.GEVmin2}
\alias{log_pdf.GEVmin2}
\title{Evaluate the probability density function of a GEVmin2 distribution}
\usage{
\method{pdf}{GEVmin2}(d, x, ...)

\method{log_pdf}{GEVmin2}(d, x, ...)
}
\arguments{
\item{d}{A `GEVmin2` object created by a call to [GEVmin2()].}

\item{x}{A vector of elements whose pdf you would like to
determine given the distribution `d`.}

\item{...}{Unused. Unevaluated arguments will generate a warning to
catch mispellings or other possible errors.}
}
\value{
A vector of pdf values, one for each element of `x`.
}
\description{
Evaluate the probability density function of a GEVmin2 distribution
}
\examples{

set.seed(27)

X <- GEVmin2(loc=2,scale=1,lbound=0)
X

random(X, 10)

pdf(X, 0.7)
log_pdf(X, 0.7)

cdf(X, 0.7)
quantile(X, 0.7)

cdf(X, quantile(X, 0.7))
quantile(X, cdf(X, 0.7))
}
