% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPD.R
\name{pdf.GPD}
\alias{pdf.GPD}
\alias{log_pdf.GPD}
\title{Evaluate the probability density function of a GEV distribution}
\usage{
\method{pdf}{GPD}(d, x, ...)

\method{log_pdf}{GPD}(d, x, ...)
}
\arguments{
\item{d}{A `GPD` object created by a call to [GPD()].}

\item{x}{A vector of elements whose pdf you would like to
determine given the distribution `d`.}

\item{...}{Unused. Unevaluated arguments will generate a warning to
catch mispellings or other possible errors.}
}
\value{
A vector of pdf values, one for each element of `x`.
}
\description{
Evaluate the probability density function of a GEV distribution
}
\examples{

set.seed(27)

X <- GPD(loc=0,scale=1,shape=0.2)
X

random(X, 10)

pdf(X, 0.7)
log_pdf(X, 0.7)

cdf(X, 0.7)
quantile(X, 0.7)

cdf(X, quantile(X, 0.7))
quantile(X, cdf(X, 0.7))
}
