% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPDmin.R
\name{pdf.GPDmin}
\alias{pdf.GPDmin}
\alias{log_pdf.GPDmin}
\title{Evaluate the probability density function of a GPDmin distribution}
\usage{
\method{pdf}{GPDmin}(d, x, ...)

\method{log_pdf}{GPDmin}(d, x, ...)
}
\arguments{
\item{d}{A `GPDmin` object created by a call to [GPDmin()].}

\item{x}{A vector of elements whose pdf you would like to
determine given the distribution `d`.}

\item{...}{Unused. Unevaluated arguments will generate a warning to
catch mispellings or other possible errors.}
}
\value{
A vector of pdf values, one for each element of `x`.
}
\description{
Evaluate the probability density function of a GPDmin distribution
}
\examples{

set.seed(27)

X <- GPDmin(loc=1,scale=1,shape=-0.2)
X

random(X, 10)

pdf(X, 0.7)
log_pdf(X, 0.7)

cdf(X, 0.7)
quantile(X, 0.7)

cdf(X, quantile(X, 0.7))
quantile(X, cdf(X, 0.7))
}
