% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpqrMethods.R
\name{pdf_trim}
\alias{pdf_trim}
\title{Pdf of a trimmed distribution}
\usage{
pdf_trim(d, x, trim = c(-Inf, Inf), doTrunc = FALSE)
}
\arguments{
\item{d}{distributions3 object}

\item{x}{Numeric vector, values at which the pdf is evaluated}

\item{trim}{Numeric vector of size 2, trimming bounds.}

\item{doTrunc}{Logical, do truncation? default FALSE, i.e. rectification is used.}
}
\value{
a numeric vector containing the pdf values.
}
\description{
Evaluate the pdf of a trimmed distribution.
Note that in the case of a rectified distribution, the
pdf at a bound is computed from the parent distribution d,
and is hence not equal to the probability of being equal to the bound.
}
\examples{
# Define Normal distribution
norm <- Normal(mu=0.75,sigma=0.5)
# Define a grid of values x
x=seq(-1,2,,100)
# Compare pdf of basic / rectified / truncated Gaussian distributions
y0=pdf_trim(norm,x)
yr=pdf_trim(norm,x,trim=c(0,1))
yt=pdf_trim(norm,x,trim=c(0,1),doTrunc=TRUE)
plot(x,y0,type='l',ylim=c(0,max(yt)))
points(x,yr,col='red')
points(x,yt,col='blue',pch=3)
}
