% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_trim}
\alias{plot_trim}
\title{Plot summary of a trimmed distribution: pdf + cdf + quantile + random}
\usage{
plot_trim(
  d,
  trim = c(-Inf, Inf),
  doTrunc = FALSE,
  xgrid = quantile_trim(d, seq(0.001, 0.999, length.out = 1000)),
  pgrid = seq(0.001, 0.999, length.out = 1000),
  nsim = 200,
  line.size = 1,
  line.colour = "black",
  area.fill = "salmon",
  area.alpha = 0.5,
  point.colour = "black",
  point.size = 2,
  lab.size = 12,
  txt.size = 4
)
}
\arguments{
\item{d}{distributions3 object}

\item{trim}{Numeric vector of size 2, trimming bounds.}

\item{doTrunc}{Logical, do truncation? default FALSE, i.e. rectification is used.}

\item{xgrid}{Numeric vector, x-values where pdf and cdf are evaluated}

\item{pgrid}{Numeric vector in (0,1), probabilities where quantile function is evaluated}

\item{nsim}{Integer, number of simulated values}

\item{line.size, line.colour, area.fill, area.alpha, point.colour, point.size, lab.size, txt.size, }{graphical parameters}
}
\value{
a ggplot object.
}
\description{
Plot summary of a trimmed distribution: pdf + cdf + quantile + random
}
\examples{
# Define Normal distribution
norm <- Normal(mu=0.75,sigma=0.5)
plot_trim(norm)
plot_trim(norm,trim=c(0,1))
plot_trim(norm,trim=c(0,1),doTrunc=TRUE)
}
