% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_trim_cdf}
\alias{plot_trim_cdf}
\title{Plot cdf of a trimmed distribution}
\usage{
plot_trim_cdf(
  d,
  trim = c(-Inf, Inf),
  doTrunc = FALSE,
  xgrid = quantile_trim(d, seq(0.001, 0.999, length.out = 1000)),
  line.size = 1,
  line.colour = "black",
  lab.size = 14
)
}
\arguments{
\item{d}{distributions3 object}

\item{trim}{Numeric vector of size 2, trimming bounds.}

\item{doTrunc}{Logical, do truncation? default FALSE, i.e. rectification is used.}

\item{xgrid}{Numeric vector, x-values where cdf is evaluated}

\item{line.size, line.colour, lab.size, }{graphical parameters}
}
\value{
a ggplot object.
}
\description{
Plot cdf of a trimmed distribution
}
\examples{
# Define Normal distribution
norm <- Normal(mu=0.75,sigma=0.5)
plot_trim_cdf(norm)
plot_trim_cdf(norm,trim=c(0,1))
plot_trim_cdf(norm,trim=c(0,1),doTrunc=TRUE)
}
