% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_trim_pdf}
\alias{plot_trim_pdf}
\title{Plot pdf of a trimmed distribution}
\usage{
plot_trim_pdf(
  d,
  trim = c(-Inf, Inf),
  doTrunc = FALSE,
  xgrid = quantile_trim(d, seq(0.001, 0.999, length.out = 1000)),
  line.size = 1,
  line.colour = "black",
  area.fill = "salmon",
  area.alpha = 0.5,
  point.colour = "black",
  point.size = 3,
  lab.size = 14,
  txt.size = 5
)
}
\arguments{
\item{d}{distributions3 object}

\item{trim}{Numeric vector of size 2, trimming bounds.}

\item{doTrunc}{Logical, do truncation? default FALSE, i.e. rectification is used.}

\item{xgrid}{Numeric vector, x-values where pdf is evaluated}

\item{line.size, line.colour, area.fill, area.alpha, point.colour, point.size, lab.size, txt.size, }{graphical parameters}
}
\value{
a ggplot object.
}
\description{
Plot pdf of a trimmed distribution
}
\examples{
# Define Normal distribution
norm <- Normal(mu=0.75,sigma=0.5)
plot_trim_pdf(norm)
plot_trim_pdf(norm,trim=c(0,1))
plot_trim_pdf(norm,trim=c(0,1),doTrunc=TRUE)
}
