% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_trim_random}
\alias{plot_trim_random}
\title{Illustrate random function of a trimmed distribution}
\usage{
plot_trim_random(
  d,
  trim = c(-Inf, Inf),
  doTrunc = FALSE,
  nsim = 200,
  point.colour = "black",
  point.size = 3,
  lab.size = 14
)
}
\arguments{
\item{d}{distributions3 object}

\item{trim}{Numeric vector of size 2, trimming bounds.}

\item{doTrunc}{Logical, do truncation? default FALSE, i.e. rectification is used.}

\item{nsim}{Integer, number of simulated values}

\item{point.size, point.colour, lab.size, }{graphical parameters}
}
\value{
a ggplot object.
}
\description{
Illustrate random function of a trimmed distribution
}
\examples{
# Define Normal distribution
norm <- Normal(mu=0.75,sigma=0.5)
plot_trim_random(norm)
plot_trim_random(norm,trim=c(0,1))
plot_trim_random(norm,trim=c(0,1),doTrunc=TRUE)
}
