% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPDmin.R
\name{quantile.GPDmin}
\alias{quantile.GPDmin}
\title{Determine quantiles of a GPDmin distribution}
\usage{
\method{quantile}{GPDmin}(d, p, ...)
}
\arguments{
\item{d}{A `GPDmin` object created by a call to [GPDmin()].}

\item{p}{A vector of probabilites.}

\item{...}{Unused. Unevaluated arguments will generate a warning to
catch mispellings or other possible errors.}
}
\value{
A vector of quantiles, one for each element of `p`.
}
\description{
`quantile()` is the inverse of `cdf()`.
}
\examples{

set.seed(27)

X <- GPDmin(loc=1,scale=1,shape=-0.2)
X

random(X, 10)

pdf(X, 0.7)
log_pdf(X, 0.7)

cdf(X, 0.7)
quantile(X, 0.7)

cdf(X, quantile(X, 0.7))
quantile(X, cdf(X, 0.7))
}
