% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPDmin2.R
\name{quantile.GPDmin2}
\alias{quantile.GPDmin2}
\title{Determine quantiles of a GPDmin2 distribution}
\usage{
\method{quantile}{GPDmin2}(d, p, ...)
}
\arguments{
\item{d}{A `GPDmin2` object created by a call to [GPDmin2()].}

\item{p}{A vector of probabilites.}

\item{...}{Unused. Unevaluated arguments will generate a warning to
catch mispellings or other possible errors.}
}
\value{
A vector of quantiles, one for each element of `p`.
}
\description{
`quantile()` is the inverse of `cdf()`.
}
\examples{

set.seed(27)

X <- GPDmin2(loc=2,scale=1,lbound=0)
X

random(X, 10)

pdf(X, 0.7)
log_pdf(X, 0.7)

cdf(X, 0.7)
quantile(X, 0.7)

cdf(X, quantile(X, 0.7))
quantile(X, cdf(X, 0.7))
}
