% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KDB.R
\name{quantile.KDB}
\alias{quantile.KDB}
\title{Determine quantiles of a KDB distribution}
\usage{
\method{quantile}{KDB}(d, p, ...)
}
\arguments{
\item{d}{A `KDB` object created by a call to [KDB()].}

\item{p}{A vector of probabilites.}

\item{...}{Unused. Unevaluated arguments will generate a warning to
catch mispellings or other possible errors.}
}
\value{
A vector of quantiles, one for each element of `p`.
}
\description{
`quantile()` is the inverse of `cdf()`.
}
\examples{

set.seed(27)

X <- KDB(alpha=2,beta=5)
X

random(X, 10)

pdf(X, 0.7)
log_pdf(X, 0.7)

cdf(X, 0.7)
quantile(X, 0.7)

cdf(X, quantile(X, 0.7))
quantile(X, cdf(X, 0.7))
}
