% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Triangular.R
\name{quantile.Triangular}
\alias{quantile.Triangular}
\title{Determine quantiles of a Triangular distribution}
\usage{
\method{quantile}{Triangular}(d, p, ...)
}
\arguments{
\item{d}{A `Triangular` object created by a call to [Triangular()].}

\item{p}{A vector of probabilites.}

\item{...}{Unused. Unevaluated arguments will generate a warning to
catch mispellings or other possible errors.}
}
\value{
A vector of quantiles, one for each element of `p`.
}
\description{
`quantile()` is the inverse of `cdf()`.
}
\examples{

set.seed(27)

X <- Triangular(a=0,b=4,c=1)
X

random(X, 10)

pdf(X, 0.7)
log_pdf(X, 0.7)

cdf(X, 0.7)
quantile(X, 0.7)

cdf(X, quantile(X, 0.7))
quantile(X, cdf(X, 0.7))
}
