% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpqrMethods.R
\name{quantile_trim}
\alias{quantile_trim}
\title{Quantile of a trimmed distribution}
\usage{
quantile_trim(d, p, trim = c(-Inf, Inf), doTrunc = FALSE)
}
\arguments{
\item{d}{distributions3 object}

\item{p}{Numeric vector in [0,1], probabilities at which quantiles are evaluated}

\item{trim}{Numeric vector of size 2, trimming bounds.}

\item{doTrunc}{Logical, do truncation? default FALSE, i.e. rectification is used.}
}
\value{
a numeric vector containing the quantiles.
}
\description{
Evaluate the quantile of a trimmed distribution.
}
\examples{
# Define Normal distribution
norm <- Normal(mu=0.75,sigma=0.5)
# Define a grid of values p
p=seq(0,1,,100)
# Compare quantiles of basic / rectified / truncated Gaussian distributions
y0=quantile_trim(norm,p)
yr=quantile_trim(norm,p,trim=c(0,1))
yt=quantile_trim(norm,p,trim=c(0,1),doTrunc=TRUE)
plot(p,y0,type='l')
points(p,yr,col='red')
points(p,yt,col='blue',pch=3)
}
