% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEVmin2.R
\name{random.GEVmin2}
\alias{random.GEVmin2}
\title{Draw a random sample from a GEVmin2 distribution}
\usage{
\method{random}{GEVmin2}(d, n = 1L, ...)
}
\arguments{
\item{d}{A `GEVmin2` object created by a call to [GEVmin2()].}

\item{n}{The number of samples to draw. Defaults to `1L`.}

\item{...}{Unused. Unevaluated arguments will generate a warning to
catch mispellings or other possible errors.}
}
\value{
A numeric vector of length `n`.
}
\description{
Draw a random sample from a GEVmin2 distribution
}
\examples{

set.seed(27)

X <- GEVmin2(loc=2,scale=1,lbound=0)
X

random(X, 10)

pdf(X, 0.7)
log_pdf(X, 0.7)

cdf(X, 0.7)
quantile(X, 0.7)

cdf(X, quantile(X, 0.7))
quantile(X, cdf(X, 0.7))
}
