% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPDmin.R
\name{random.GPDmin}
\alias{random.GPDmin}
\title{Draw a random sample from a GPDmin distribution}
\usage{
\method{random}{GPDmin}(d, n = 1L, ...)
}
\arguments{
\item{d}{A `GPDmin` object created by a call to [GPDmin()].}

\item{n}{The number of samples to draw. Defaults to `1L`.}

\item{...}{Unused. Unevaluated arguments will generate a warning to
catch mispellings or other possible errors.}
}
\value{
A numeric vector of length `n`.
}
\description{
Draw a random sample from a GPDmin distribution
}
\examples{

set.seed(27)

X <- GPDmin(loc=1,scale=1,shape=-0.2)
X

random(X, 10)

pdf(X, 0.7)
log_pdf(X, 0.7)

cdf(X, 0.7)
quantile(X, 0.7)

cdf(X, quantile(X, 0.7))
quantile(X, cdf(X, 0.7))
}
