% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KDB4.R
\name{random.KDB4}
\alias{random.KDB4}
\title{Draw a random sample from a KDB4 distribution}
\usage{
\method{random}{KDB4}(d, n = 1L, ...)
}
\arguments{
\item{d}{A `KDB4` object created by a call to [KDB4()].}

\item{n}{The number of samples to draw. Defaults to `1L`.}

\item{...}{Unused. Unevaluated arguments will generate a warning to
catch mispellings or other possible errors.}
}
\value{
A numeric vector of length `n`.
}
\description{
Draw a random sample from a KDB4 distribution
}
\examples{

set.seed(27)

X <- KDB4(alpha=2,beta=5,a=0.5,b=3)
X

random(X, 10)

pdf(X, 0.7)
log_pdf(X, 0.7)

cdf(X, 0.7)
quantile(X, 0.7)

cdf(X, quantile(X, 0.7))
quantile(X, cdf(X, 0.7))
}
