% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Triangular.R
\name{random.Triangular}
\alias{random.Triangular}
\title{Draw a random sample from a Triangular distribution}
\usage{
\method{random}{Triangular}(d, n = 1L, ...)
}
\arguments{
\item{d}{A `Triangular` object created by a call to [Triangular()].}

\item{n}{The number of samples to draw. Defaults to `1L`.}

\item{...}{Unused. Unevaluated arguments will generate a warning to
catch mispellings or other possible errors.}
}
\value{
A numeric vector of length `n`.
}
\description{
Draw a random sample from a Triangular distribution
}
\examples{

set.seed(27)

X <- Triangular(a=0,b=4,c=1)
X

random(X, 10)

pdf(X, 0.7)
log_pdf(X, 0.7)

cdf(X, 0.7)
quantile(X, 0.7)

cdf(X, quantile(X, 0.7))
quantile(X, cdf(X, 0.7))
}
