% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Triangular.R
\name{Triangular}
\alias{Triangular}
\title{Create a Triangular distribution}
\usage{
Triangular(a = 0, b = 1, c = 0.5)
}
\arguments{
\item{a}{The a parameter (lower bound). `a` can be any value in the set of real
numbers. Defaults to `0`.}

\item{b}{The b parameter (upper bound). `b` can be any value in the set of real
numbers. It should be strictly bigger than `a`. Defaults to `1`.}

\item{c}{The c parameter (peak). `c` can be any value in the set of real
numbers. It should be strictly between `a` and `b`. Defaults to `0.5`.}
}
\value{
A `Triangular` object.
}
\description{
Create a Triangular distribution
}
\examples{

set.seed(27)

X <- Triangular(a=0,b=4,c=1)
X

random(X, 10)

pdf(X, 0.7)
log_pdf(X, 0.7)

cdf(X, 0.7)
quantile(X, 0.7)

cdf(X, quantile(X, 0.7))
quantile(X, cdf(X, 0.7))
}
\seealso{
Other continuous distributions: 
\code{\link{GEVmin2}()},
\code{\link{GEVmin}()},
\code{\link{GEV}()},
\code{\link{GPDmin2}()},
\code{\link{GPDmin}()},
\code{\link{GPD}()},
\code{\link{KDB4}()},
\code{\link{KDB}()}
}
\concept{continuous distributions}
