% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{trimLL}
\alias{trimLL}
\title{Log-likelihood for a trimmed distribution}
\usage{
trimLL(param, d, y, trim, doTrunc, par.low, par.high)
}
\arguments{
\item{param}{Numeric vector, parameters}

\item{d}{distributions3 object, assumed distribution.}

\item{y}{Numeric vector, data}

\item{trim}{Numeric vector of size 2, trimming bounds.}

\item{doTrunc}{Logical, do truncation? default FALSE, i.e. rectification is used.}

\item{par.low}{Numeric vector of size nPar(d), lower bounds for parameters.}

\item{par.high}{Numeric vector of size nPar(d), higher bounds for parameters.}
}
\value{
The log-likelihood (may return NA's or -Inf)
}
\description{
Log-likelihood for a trimmed distribution
}
\keyword{internal}
